[APPC-2240] - Move SkuButtonsViewHolder to separate class.
