+++ app/src/main/java/com/asfoundation/wallet/promotions/voucher/SkuButtonsAdapter.kt
@@ -59,35 +59,6 @@ class SkuButtonsAdapter(
   fun getSelectedSku(): SkuButtonModel = buttonModels.get(selectedPosition)
 }
 
-class SkuButtonsViewHolder(private val button: Button) : RecyclerView.ViewHolder(button) {
-
-  fun bind(position: Int, selectedPosition: Int,
-           skuButtonModel: SkuButtonModel,
-           activatedButton: AtomicReference<Button?>,
-           onSkuClick: PublishSubject<Int>
-  ) {
-    button.text = skuButtonModel.title
-
-    if (selectedPosition == position) {
-      button.isActivated = true
-    } else {
-      button.isActivated = false
-    }
-
-    button.setOnClickListener {
-      if (activatedButton.get() == null) {
-        activatedButton.set(button)
-      } else {
-        activatedButton.get()
-            ?.setActivated(false)
-        activatedButton.set(button)
-      }
-
-      onSkuClick.onNext(position)
-    }
-  }
-}
-
 class MarginItemDecoration(private val spaceSize: Int) : RecyclerView.ItemDecoration() {
   override fun getItemOffsets(
       outRect: Rect, view: View,
